refactor(core): move task_enqueue into core/models beside its siblings
The flat-core ratchet (tests/quality/test_no_flat_core_regrowth.py) read 110
leaves against a ceiling of 109: the new #4085 seam landed at the core root
rather than in the subpackage that owns its concern. task_repair.py and
task_handoff.py already live under core/models/ for the same reason -
core/models/task.py is past the 500-LOC cap, so its Task-lifecycle concerns
are split one leaf each.

Pure relocation: the module body is unchanged, the mirror test moves with it
(t3 tool test-path-mirror resolves the new package), and the three import
sites plus the Sphinx cross-reference and the BLUEPRINT path string follow.
No re-export, no shim at the old path, no tach entry - both consumers are
interface -> domain, which layered mode already permits.
