+++ src/teatree/core/management/commands/tasks.py
@@ -19,9 +19,9 @@ from teatree.core.management.commands.tasks_session_view import (
 )
 from teatree.core.modelkit.task_failure_taxonomy import CANCELLED_PREFIX, is_environmental
 from teatree.core.models import InvalidTransitionError, Task, TaskAttempt, Ticket
+from teatree.core.models.task_enqueue import TaskEnqueueError, enqueue_phase_task
 from teatree.core.overlay_loader import get_overlay_for_ticket
 from teatree.core.session_identity import current_session_id
-from teatree.core.task_enqueue import TaskEnqueueError, enqueue_phase_task
 
 logger = logging.getLogger(__name__)
